+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### 2.9.0
+
+- Add filters to the Bird's Eye Report
+- Add team and domain hierarchies to the Bird's Eye Report
+- Add option to group by domain to the Bird's Eye Report
+
 ### 2.8.0
 
 - Display Initiatives notification schedule